+++ src/main/java/fr/estecka/variantscit/modules/cache/CacheBuilder.java
@@ -33,22 +33,22 @@ public abstract class CacheBuilder
 		modules.entrySet().stream()
 			.sorted(Comparator.comparing(Map.Entry::getKey))
 			.forEach(e->{
-				list.addAll(BuildList(e.getValue()));
+				list.addAll(BuildPriorityGroup(e.getValue()));
 			});
 
 		return list.UnwrapIfSingle();
 	}
 
-	static public List<IBakedModule> BuildList(List<IBakedModule> modules){
+	static public List<IBakedModule> BuildPriorityGroup(List<IBakedModule> modules){
 		List<IBakedModule> result = new ArrayList<>(modules);
 
+		// Create group caches for modules with the same fingerprints.
 		for (int ia=0; ia<result.size(); ++ia)
 		for (int ib=result.size()-1; ia<ib; --ib)
 		{
 			IBakedModule a = result.get(ia);
 			IBakedModule b = result.get(ib);
 
-			// Create group caches for modules with the same fingerprints.
 			if (a.GetCacheKeys().CompareTo(b.GetCacheKeys()).isEqual()){
 				result.set(ia, CacheModules(a,b));
 				result.remove(ib);
@@ -66,7 +66,7 @@ public abstract class CacheBuilder
 	}
 
 	/**
-	 * Creates a single cache encompassing contains all provided modules, whilst
+	 * Creates  a single  cache  that encompasses  all provided  modules, whilst
 	 * minimizing the amount of wrap layers.
 	 */
 	static public CacheModule CacheModules(IBakedModule... modules){